+++ src/repoe/__init__.py
@@ -1,1 +1,2 @@
-from .functions import get_class_for_base
+from .base import get_class_for_base, get_bases
+from .mod import get_mods