+++ tests/test_utilities/__init__.py
@@ -1,4 +1,4 @@
-from .miscellaneous import create_filter
+from .miscellaneous import create_filter, create_sieve_for_text, create_sieve_for_pattern
 from .function_mock import FunctionMock
 from .file_mock import FileMock
 from .web_get_mock import WebGetMock