+++ changelog.md
@@ -3,6 +3,7 @@
 * Added **Omens** and **Tattoos** to `.econ`.
 * Fixed a bug where empty strings (`""`) could not count as admissible values in filter lines.
 * Added a caching system to significantly improve performance for web requests.
+* Added a `.game` handler, which allows pulling game data into the filter. More information can be found on the [`.game` wiki page](https://github.com/ajoscram/PoE-Filter-Generator/wiki/.game).
 * Added a `:clean` command which allows for manual deletion of the cache. More information can be found on the [`:clean` wiki page](https://github.com/ajoscram/PoE-Filter-Generator/wiki/:clean).
 * The `:help` command now displays a list of the terms that can be queried with it after displaying [the Usage article](https://github.com/ajoscram/PoE-Filter-Generator/wiki/Usage).
 * The `:path` command now supports an argument called `remove`, which removes the directory where the PFG executable resides from the PATH environment variable.